refactor: Update NavBar component to handle dynamic rendering of buttons
The NavBar component in `src/components/navbar/NavBar.tsx` has been updated to handle dynamic rendering of buttons based on the current pathname. The `usePathname` hook from the `next/navigation` package is now used to get the current pathname. If the pathname matches `AppPath.add`, a "Back" button is rendered that navigates to the previous page. Otherwise, a "New" button is rendered that navigates to the `/add` page. This refactor improves the code by providing a more flexible and intuitive navigation experience for the user.